Move classes to core directory
I want "handlers" directory contain only base handlers
