+++ src/core/HandlerRegistry.ts
