+++ src/services/handlers.ts
@@ -1,4 +1,4 @@
-import { HandlerRegistry } from "../handlers/HandlerRegistry";
+import { HandlerRegistry } from "../core/HandlerRegistry";
 import logger from "./logger";
 
 export default new HandlerRegistry(logger);
